refactor; move view iterator
