+++ cognite/neat/_graph/extractors/_views.py
@@ -1,12 +1,15 @@
 import typing
-from collections.abc import Iterable, Set
+from collections.abc import Iterable, Iterator, Set
+from functools import cached_property
 
+from cognite.client import data_modeling as dm
 from cognite.client.data_classes.data_modeling import ViewId, ViewIdentifier
 from cognite.client.data_classes.data_modeling.instances import Instance
 from cognite.client.utils.useful_types import SequenceNotStr
 from rdflib import Namespace
 
 from cognite.neat._client import NeatClient
+from cognite.neat._constants import is_readonly_property
 from cognite.neat._issues.errors import ResourceRetrievalError
 from cognite.neat._shared import Triple
 
@@ -83,9 +86,9 @@ class ViewExtractor(BaseExtractor):
         retrieved_list = client.data_modeling.views.retrieve(view_id)
         if not retrieved_list:
             raise ResourceRetrievalError(ViewId.load(view_id), "view", "View is missing in CDF")
-        latest_view = retrieved_list.latest_version()
+        latest_view = max(retrieved_list, key=lambda v: v.last_updated_time)
         instance_iterator = _ViewInstanceIterator(client, latest_view, instance_space)
-        total = instance_iterator.total
+        total = instance_iterator.count
         return cls(
             latest_view.as_id(),
             instance_iterator,
@@ -113,3 +116,82 @@ class ViewExtractor(BaseExtractor):
             edge_type=self.edge_type,
         )
         yield from instance_iterator.extract()
+
+
+class _ViewInstanceIterator(Iterable[Instance]):
+    def __init__(self, client: NeatClient, view: dm.View, instance_space: str | SequenceNotStr[str] | None = None):
+        self.client = client
+        self.view = view
+        self.instance_space = instance_space
+
+    @cached_property
+    def count(self) -> int:
+        node_count = edge_count = 0
+        if self.view.used_for in ("node", "all"):
+            node_result = self.client.data_modeling.instances.aggregate(
+                view=self.view.as_id(),
+                aggregates=dm.aggregations.Count("externalId"),
+                instance_type="node",
+                space=self.instance_space,
+            ).value
+            if node_result:
+                node_count = int(node_result)
+        if self.view.used_for in ("edge", "all"):
+            edge_result = self.client.data_modeling.instances.aggregate(
+                view=self.view.as_id(),
+                aggregates=dm.aggregations.Count("externalId"),
+                instance_type="edge",
+                space=self.instance_space,
+            ).value
+            if edge_result:
+                edge_count = int(edge_result)
+        return node_count + edge_count
+
+    def __iter__(self) -> Iterator[Instance]:
+        view_id = self.view.as_id()
+        read_only_properties = {
+            prop_id
+            for prop_id, prop in self.view.properties.items()
+            if isinstance(prop, dm.MappedProperty)
+            and is_readonly_property(prop.container, prop.container_property_identifier)
+        }
+        # All nodes and edges with properties
+        if self.view.used_for in ("node", "all"):
+            node_iterable: Iterable[Instance] = self.client.instances.iterate(
+                instance_type="node",
+                source=view_id,
+                space=self.instance_space,
+            )
+            if read_only_properties:
+                node_iterable = self._remove_read_only_properties(node_iterable, read_only_properties, view_id)
+            yield from node_iterable
+
+        if self.view.used_for in ("edge", "all"):
+            yield from self.client.instances.iterate(
+                instance_type="edge",
+                source=view_id,
+                space=self.instance_space,
+            )
+
+        for prop in self.view.properties.values():
+            if isinstance(prop, dm.EdgeConnection):
+                if prop.edge_source:
+                    # All edges with properties are extracted from the edge source
+                    continue
+                yield from self.client.instances.iterate(
+                    instance_type="edge",
+                    filter_=dm.filters.Equals(
+                        ["edge", "type"], {"space": prop.type.space, "externalId": prop.type.external_id}
+                    ),
+                    space=self.instance_space,
+                )
+
+    @staticmethod
+    def _remove_read_only_properties(
+        nodes: Iterable[Instance], read_only_properties: Set[str], view_id: dm.ViewId
+    ) -> Iterable[Instance]:
+        for node in nodes:
+            if properties := node.properties.get(view_id):
+                for read_only in read_only_properties:
+                    properties.pop(read_only, None)
+            yield node
